+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Apache.Ignite.Core.Tests.csproj
@@ -116,14 +116,14 @@
     <Compile Include="Cache\CacheTimeoutOnPmeTransactionalTest.cs" />
     <Compile Include="Cache\DataRegionMetricsTest.cs" />
     <Compile Include="Cache\DataStorageMetricsTest.cs" />
-    <Compile Include="Cache\Near\CacheNearCreateDestroy.cs" />
-    <Compile Include="Cache\Near\CacheNearTopologyChangeTest.cs" />
-    <Compile Include="Cache\Near\CacheNearTest.cs" />
-    <Compile Include="Cache\Near\FailingCacheStore.cs" />
-    <Compile Include="Cache\Near\Foo.cs" />
-    <Compile Include="Cache\Near\ScanQueryNearCacheFilter.cs" />
-    <Compile Include="Cache\Near\ScanQueryNoNearCacheFilter.cs" />
-    <Compile Include="Cache\Near\StoreNoNearCacheFilter.cs" />
+    <Compile Include="Cache\Platform\CacheNearCreateDestroy.cs" />
+    <Compile Include="Cache\Platform\CacheNearTopologyChangeTest.cs" />
+    <Compile Include="Cache\Platform\CacheNearTest.cs" />
+    <Compile Include="Cache\Platform\FailingCacheStore.cs" />
+    <Compile Include="Cache\Platform\Foo.cs" />
+    <Compile Include="Cache\Platform\ScanQueryNearCacheFilter.cs" />
+    <Compile Include="Cache\Platform\ScanQueryNoNearCacheFilter.cs" />
+    <Compile Include="Cache\Platform\StoreNoNearCacheFilter.cs" />
     <Compile Include="Cache\PartitionPreloadTest.cs" />
     <Compile Include="Cache\PersistenceTest.cs" />
     <Compile Include="Cache\PersistentStoreTestObsolete.cs" />
