+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Cache/Platform/CacheNearTest.cs
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.Core.Tests.Cache.Near
+namespace Apache.Ignite.Core.Tests.Cache.Platform
 {
     using System;
     using System.Collections.Generic;
