+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Cache/Platform/ScanQueryNearCacheFilter.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * Copyright 2019 GridGain Systems, Inc. and Contributors.
  *
  * Licensed under the GridGain Community Edition License (the "License");
@@ -14,35 +14,48 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.Core.Tests.Cache.Near
+namespace Apache.Ignite.Core.Tests.Cache.Platform
 {
+    using System.Security;
     using Apache.Ignite.Core.Cache;
     using Apache.Ignite.Core.Resource;
     using NUnit.Framework;
 
     /// <summary>
-    /// Entry filter for cache store: ensures that values do not come from Platform Near Cache.
+    /// Scan query filter that checks whether values come from native near cache.
     /// </summary>
-    public class StoreNoNearCacheFilter : ICacheEntryFilter<int, Foo>
+    public class ScanQueryNearCacheFilter : ICacheEntryFilter<int, Foo>
     {
         /// <summary>
         /// Gets or sets the cache name.
         /// </summary>
         public string CacheName { get; set; }
         
+        /// <summary>
+        /// Gets or sets the key that should cause an exception in <see cref="Invoke"/>. 
+        /// </summary>
+        public int? FailKey { get; set; }
+        
+        /// <summary>
+        /// Injected Ignite.
+        /// </summary>
         [InstanceResource]
         public IIgnite Ignite { get; set; }
         
+        /** <inheritdoc /> */
         public bool Invoke(ICacheEntry<int, Foo> entry)
         {
+            if (entry.Key == FailKey)
+            {
+                throw new SecurityException("Crash in filter");
+            }
+            
             var cache = Ignite.GetCache<int, Foo>(CacheName);
+            var nearVal = cache.LocalPeek(entry.Key, CachePeekMode.Platform);
 
-            var nearVal = cache.LocalPeek(entry.Key);
+            Assert.AreSame(nearVal, entry.Value);
 
-            Assert.AreNotSame(nearVal, entry.Value);
-            Assert.AreNotEqual(nearVal.Bar, entry.Value.Bar);
-
-            return Ignite.GetAffinity(CacheName).IsPrimary(Ignite.GetCluster().GetLocalNode(), entry.Key);
+            return true;
         }
     }
 }
